Improve mutator for multiple types
