+++ src/mutator/mod.rs
@@ -1,2 +1,4 @@
 mod magic_values;
-pub mod mutator;
+pub mod mutator_felt252;
+//pub mod mutator_u8;
+//mod mutator;
