+++ tests1.0/fuzzinglabs.cairo
@@ -18,12 +18,12 @@ mod test_contract {
 ref self: ContractState,
     f: felt252,
     u: felt252,
-    z: felt252,
-    z2: felt252,
-    i: felt252,
-    n: felt252,
-    g: felt252,
-    l: felt252,
+    z: u16,
+    z2: u32,
+    i: u64,
+    n: u128,
+    g: u128,
+    l: u256,
     a: felt252,
     b: felt252,
     s: u8,
